requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
inflection==0.5.1
jsonschema==4.17.3
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
